Compare the FITS float as the value the core reports

+++ crates/core/tests/fits_real.rs
@@ -58,7 +58,7 @@ fn a_real_binary_tables_columns_are_typed_by_its_header() {
     assert!(matches!(&c2.value, Value::Str(s) if s.trim() == "abc"), "{:?}", c2.value);
 
     // `TFORM3 = '1E'`: one float.
-    assert_eq!(ev.node(&doc, &at(&[2, 0, 0, 2, 0])).unwrap().value, Value::Float(1.1f32 as f64));
+    assert_eq!(ev.node(&doc, &at(&[2, 0, 0, 2, 0])).unwrap().value, Value::Float(1.1));
 
     // `TFORM4 = '1L'`: a logical, written as the letter T or F; the first row says F.
     let c4 = ev.node(&doc, &at(&[2, 0, 0, 3])).unwrap();
